Move the main package to the internal packages
